Update NewNVDNistScraper.py
The updates to the nvdlib Python library broke the script.  The updates to this script fix the issues that were caused by the updates. The results are now outputted to a text file that is formatted in a way that it will allow it to be uploaded to an Excel spreadsheet, making the data manageable and workable! 

Vulnerabilities will now be able to be sorted!